requirements.txt
asgiref==3.10.0
Django==5.2.8
django-modeltranslation==0.19.17
pillow==12.0.0
psycopg2-binary==2.9.11
six==1.17.0
sqlparse==0.5.3
